return a double instead of a Long.

+++ android/src/main/java/com/sudoplz/reactnativeamplitudeanalytics/RNAmplitudeSDK.java
@@ -76,8 +76,8 @@ public class RNAmplitudeSDK extends ReactContextBaseJavaModule {
 
   @ReactMethod
   public void getSessionId(Promise promise) {
-    long sessionId = Amplitude.getInstance().getSessionId();
-    promise.resolve(sessionId);
+    Long sessionId = Amplitude.getInstance().getSessionId();
+    promise.resolve(sessionId.doubleValue());
   }
 
   @ReactMethod
